Improve usage of NEO web service API
Calls to the NEO API are now defining start and end date explictly.
Start date is set to the current day, end date is 7 days in advance.

Before this change the NEO API was called without explictly defining parameters
for start and end date. It was relying on undocumented default
behaviour.